Added ggplotly compatibility! (DBDimPlot-only)
Added compatibility for using the data.hover feature of the plotly package for the DBDimPlot function.

The plotly package is now a suggested dependency.

Adjusted the shape-by and lettering options of the same functions in order to adjust what happens when conflicting options are used.  (With my implementation at least, lettering is incompatibile with both shape-by and hover.)