Added a new DBHeatmap function and a relative.to.max option for gene-expression scaling
Added a new DBHeatmap function

Also added a "relative.to.max" option for gene-expression scaling that works for all functions that call on gene().

Small tweak to meta() function = it can now handle metadata names that have spaces in them!  The rest of my function may still not handle that, however, so I still advice against names with spaces.