Added new ScatterPlot function
DittoScatterPlot is a totally new function that allows making gene-gene, metadata-metadata, gene-meta scatterplots where cells/samples can also be colored or shaped based on gene or metadata.

This function handles the basic underpinnings of DBDimPlot, so I will soon be upgrading DBDimPlot to work from a DittoScatterPlot. and changing its name to DittoDimePlot.