Fixed an error in ImportDemux2Seurat
Fixed an issue stemming from use of object versus "object" that led to an error:
"trying to get slot "meta.data" from an object of a basic class ("character") with no slots"

By requesting that the user provide the quoted object name. The function now obtains the full object internally.