Adjusted multi_dittoPlot default for xlab
Added `xlab` as input for multi_dittoPlot and set its default to `NULL`.

Also added implementation of the add.title variable, which existed as an input, but was previously in the actual code.

Added coord_cartesian(clip = "off") to ridgeplotting code so that the upper bounds of the top plot are nolonger cut!