Minor bug fix: dittoBarPlot
Fixed an issue where `max`, the y-axis zoom maximum, default was filled in incorrectly for both `scale = "percent"` and `scale = "count"` methods. When not provided by the user, it now gets properly set to 1 (percent) or the max of the total counts per grouping (count).

`min`, the y-axis zoom minimum, default was changed from NULL to 0 as this is the proper default for most use cases.