Scrapped RNAseq-class for SingleCellExperiment
Reconstructed the importDESeq2 and importDGEList functions as importDittoBulk. This function will convert DGEList bulk data into a SummarizedExperiment (SE) object.  DESeqDataSets are already instances of the SE-class.  Then, the SE is converted into a SingleCellExperiment (SCE; because of it's convenient internal storage of dimensionality reductions and internal metadata && because dittoSeq already knows how to handle these!) with a couple of additional internal metadata pieces.

addDimReduction, and by extension addPrcomp, were updated to utilize the SCE `reducedDim<-` function.

All meta setter functions, previously made for the now scrapped RNAseq-class, were removed in favor of the setters already available for SCE/SEs.