Condense letter'ing, labelling, and ellipse'ing code
De-duplicates some code that was in both dittoScatterHex & dittoScatterPlot by:

Replaces the original code with an internal function, `.add_letters_ellipses_labels_if_discrete()` for handling all three of these features.  The function also handles the ignorance of such features, with a message, when the data is not discrete and is therefore incompatible.